+++ CHANGELOG.md
@@ -28,6 +28,8 @@
   top.
 - [Context menu for connections][14325].
 - [Warnings and Errors no longer become transparent][14388]
+- [GeoMap visualization is now working without need of Mapbox Token in
+  environment][14429]
 
 [13685]: https://github.com/enso-org/enso/pull/13685
 [13658]: https://github.com/enso-org/enso/pull/13658
@@ -51,6 +53,7 @@
 [14267]: https://github.com/enso-org/enso/pull/14267
 [14325]: https://github.com/enso-org/enso/pull/14325
 [14388]: https://github.com/enso-org/enso/pull/14388
+[14429]: https://github.com/enso-org/enso/pull/14429
 
 #### Enso Standard Library